Dedicated module for AWS helper data structures (#47)
* Separate AWS helper data structures into a dedicated module

* Fix asserts for exceptions in tests

* Add missing type hinting

* Replace str calls with repr
